vid_synth/requirements.txt
python-dotenv
requests
playwright
supabase
google-generativeai
gTTS
